requirements.txt
cloudpickle
gitpython
gym~=0.18.0
hydra-core
hydra-colorlog
matplotlib~=3.4.2
numba
numpy~=1.20.3
numpy-quaternion
omegaconf~=2.1.0
opencv-python~=4.5.2.54
pybullet~=3.1.7
scipy~=1.4.1
pandas~=1.2.4
tacto~=0.0.2

pytest~=6.2.4
torch~=1.8.1+cpu
open3d~=0.13.0
tqdm~=4.60.0
setuptools~=56.2.0